refactoring logic around prime number filtering + adding client side validation for empty input
