+++ src/server/app.js
@@ -1,42 +1,16 @@
-const isValidInput = (start, end) => {
-	if (
-		!isNumeric(start) ||
-		!isFinite(start) ||
-		!isNumeric(end) ||
-		!isFinite(end)
-	)
-		throw new Error(
-			'Invalid input types, please enter numeric values only'
-		);
-
-	if (!isGreaterThanOne(start) || !isGreaterThanOne(end))
-		throw new Error(
-			'Invalid input types, please enter numeric values greater than one only'
-		);
-
-	if (start > end)
-		throw new Error('Start value must be greater than end value');
-
-	return true;
-};
-
-const isNumeric = num => {
-	return typeof num === 'number';
-};
-
-const isGreaterThanOne = num => {
-	return num > 1;
-};
+import { isValidInput } from './validations';
 
 const isPrime = num => {
-	let count = 0;
-	for (let i = 1; i <= num; i++) {
-		if (num % i === 0) {
-			count++;
-		}
-		if (count > 2) {
-			return false;
-		}
+	if (num === 2) {
+		return true;
+	}
+
+	if (num % 2 === 0) {
+		return false;
+	}
+
+	for (let i = 3; i <= Math.round(Math.sqrt(num)); i++) {
+		if (num % i === 0) return false;
 	}
 
 	return true;
